Adds LineBreakSuggester
